pyproject.toml
[project]
name = "mcp-gsuite"
version = "0.4.0"
description = "MCP Server to connect to Google G-Suite"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "beautifulsoup4>=4.12.3",
 "google-api-python-client>=2.154.0",
 "httplib2>=0.22.0",
 "mcp>=1.3.0",
 "oauth2client==4.1.3",
 "python-dotenv>=1.0.1",
 "pytz>=2024.2",
 "requests>=2.32.3",
]
[[project.authors]]
name = "Markus Pfundstein"
email = "markus@life-electronic.nl"

[dependency-groups]
dev = [
    "pyright>=1.1.389",
]